latest

+++ FourierShift_MakeFile/core/fourier_filter.c
@@ -14,10 +14,10 @@ int wavenum(const int i, const int n) { return (i + n / 2) % n - n / 2; }
  * @param H_size, size of input array.
  * @param M, Size of integral.
  */
-void create_filter(double complex *H, const int H_size, const int M) {
+void create_filter(float complex *H, const int H_size, const int M) {
   for (int i = 0; i < H_size; i++) {
     int k = wavenum(i, H_size);
-    double a = 2 * M_PI / H_size * k * M;
+    float a = 2 * M_PI / H_size * k * M;
 
     if (a != M_PI / M || a != -(M_PI / M)) {
       H[i] += -(2 * pow(M, 2) * a * sin(M * a)) /
@@ -40,9 +40,8 @@ void create_filter(double complex *H, const int H_size, const int M) {
  * @param H_size, size of the filter.
  * @param shift, amount to shift.
  */
-void create_phase_shift(double complex *L, const int H_size,
-                        const double shift) {
-  double delta = remainder(shift, 1.0);
+void create_phase_shift(float complex *L, const int H_size, const float shift) {
+  float delta = remainder(shift, 1.0);
   for (int i = 0; i < H_size; i++) {
     L[i] = cexp(-2 * I * M_PI * delta * wavenum(i, H_size) / H_size);
   }
@@ -57,16 +56,16 @@ void create_phase_shift(double complex *L, const int H_size,
  * This function mulitplies the two arrays H and L to shift the H filter and
  * then peroforms normal inverse fourier transform on Z and store in z.
  */
-void shift_filter(double complex *H, double complex *L, double complex *z,
+void shift_filter(float complex *H, float complex *L, float complex *z,
                   const int H_size, const int M) {
-  double complex *Z = calloc(H_size, sizeof(double complex));
+  float complex *Z = calloc(H_size, sizeof(float complex));
   for (int i = 0; i < H_size; i++) {
     Z[i] = H[i] * L[i];
   }
 
   // Inverse Fourier Transform
   for (int i = 0; i < H_size; i++) {
-    double complex X = 0;
+    float complex X = 0;
     for (int j = 0; j < H_size; j++) {
       X += Z[j] * cexp(I * 2 * M_PI * wavenum(i, H_size) * j / H_size);
     }
@@ -81,14 +80,14 @@ void shift_filter(double complex *H, double complex *L, double complex *z,
   int n2 = H_size / 2;  // half of vector length
 
   for (int i = 0; i < n2; i++) {
-    double complex tmp = z[i];
+    float complex tmp = z[i];
     z[i] = z[i + n2];
     z[i + n2] = tmp;
   }
 
   if (H_size & 1)  // odd n, shift the rest
   {
-    double complex tmp = z[H_size - 1];
+    float complex tmp = z[H_size - 1];
     z[H_size - 1] = z[0];
     for (int i = 1; i < n2; i++) {
       z[i - 1] = z[i];
