+++ FourierShift_MakeFile/core/fourier_filter.h
@@ -11,15 +11,14 @@ int wavenum(const int i, const int n);
  * @param H_size, size of input array.
  * @param M, Size of integral.
  */
-void create_filter(double complex *H, const int H_size, const int M);
+void create_filter(float complex *H, const int H_size, const int M);
 
 /**
  * @param L, input array for store the phase shift.
  * @param H_size, size of the filter.
  * @param shift, amount to shift.
  */
-void create_phase_shift(double complex *L, const int H_size,
-                        const double shift);
+void create_phase_shift(float complex *L, const int H_size, const float shift);
 
 /**
  * @param H, original filter.
@@ -30,5 +29,5 @@ void create_phase_shift(double complex *L, const int H_size,
  * This function mulitplies the two arrays H and L to shift the H filter and
  * then peroforms normal inverse fourier transform on Z and store in z.
  */
-void shift_filter(double complex *H, double complex *L, double complex *z,
+void shift_filter(float complex *H, float complex *L, float complex *z,
                   const int H_size, const int M);
